requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
scipy>=1.11.0
